MaP_Testing.py
Acesses the MaP Toilets website and scrapes for search results. Currently only has the ability to search for one search out of the list of search results. Was trying to figure out two things. One being how to replicate a request to make the website load more search results and second being how to arrange the data frame so that the specification values appear all in one row.